Added login functionality, including login, account creation, bad info, account already exists, and direction to home page
Created a hash class with static methods to assist in hashing passwords.
Changed User class to be created with a hash and not a password, and to
not add user to database. Set up context listener to create new database
connection on context creation, and to store db in context. Set
httpsession to store User object for each user's session.