Filled in toHTML() methods for the question classes.
Have not tested yet. I laid it out so that the question html does not
include a form, otherwise we will have to have a separate form and
submit button for each question. To allow multiple questions on one page
there should be just one form, so the page html can include the form
tags and just insert the question html inside.

For the name of the question I set it it to the question number. This
should allow us to easily allow questions to answers for grading,
without having to pass hidden fields.